bugfix: Post audit report fixes
- Removed outdated dependencies such as `jup-ag` and updated versions for several packages in `Cargo.lock`.
- Enhanced `ConfigValidator` to include warnings for security risks related to token extensions and authentication configurations.
- Updated fee payer policies to default to more secure settings, ensuring better protection against unauthorized transfers.
- Refactored several structs to implement `Default` for improved initialization.
- Adjusted transaction fee response types for consistency and accuracy in handling fees.
- Use of fixed point decimals for price manipulation
- Constant equal for API KEY comparaison

+++ crates/lib/src/fee/fee.rs
@@ -316,7 +316,7 @@ impl FeeConfigUtil {
         rpc_client: &RpcClient,
         fee_in_lamports: u64,
         fee_token: Option<&str>,
-    ) -> Result<Option<f64>, KoraError> {
+    ) -> Result<Option<u64>, KoraError> {
         if let Some(fee_token) = fee_token {
             let token_mint = Pubkey::from_str(fee_token).map_err(|_| {
                 KoraError::InvalidTransaction("Invalid fee token mint address".to_string())
